Resolve certifications sometimes blank
Movies have a releases() function that provides a list of countries with information about the release in that country.
Previously the code was taking the first country in the list and using that certification. For new films this
turned out to be blank quite a lot. Code now looks through this list for any non blank entries, giving preference
to the US.